Proof-of-concept for hiding from Latest feed
This is to sort of "mute" folks who we follow and have available in the
Feed Inbox, but we don't want polluting our main feed. The biggest use
case for this is when you follow someone who posts a _lot_ and you want
to read their posts at your leisure without their posts completely
taking over the Latest feed.